moved config and option flow to config_flow.py for hassfest validation
