+++ custom_components/spotcast/__init__.py
@@ -21,9 +21,7 @@ from custom_components.spotcast.services import ServiceHandler
 from custom_components.spotcast.services.const import SERVICE_SCHEMAS
 from custom_components.spotcast.sessions.exceptions import TokenRefreshError
 from custom_components.spotcast.websocket import async_setup_websocket
-from custom_components.spotcast.config_flow.option_flow_handler import (
-    DEFAULT_OPTIONS
-)
+from custom_components.spotcast.config_flow import DEFAULT_OPTIONS
 
 __version__ = "5.0.0-b16"
 
@@ -44,7 +42,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """
 
     # ensure default options
-    updated_options = {**DEFAULT_OPTIONS, **entry.options}
+    updated_options = DEFAULT_OPTIONS | entry.options
 
     if updated_options != entry.options:
         hass.config_entries.async_update_entry(entry, options=updated_options)
