--- custom_components/spotcast/config_flow/__init__.py
@@ -1,5 +1,0 @@
-"""Module managin the config and option flow"""
-
-from custom_components.spotcast.config_flow.config_flow_handler import (
-    SpotcastFlowHandler
-)
